Feat(#5): Kafka 기능 구현
