+++ docker/docker-compose.local.yml
@@ -1,0 +1,25 @@
+services:
+  notification-server:
+    container_name: wypl-local-notification-server
+    build:
+      context: ./..
+      dockerfile: Dockerfile
+    ports:
+      - "8080:8080"
+    environment:
+      - PROFILE=local
+  kafka:
+    image: bitnami/kafka
+    container_name: kafka-server
+    ports:
+      - "9094:9094"
+    environment:
+      - KAFKA_ENABLE_KRAFT=yes
+      - KAFKA_CFG_NODE_ID=0
+      - KAFKA_CFG_PROCESS_ROLES=broker,controller
+      - KAFKA_CFG_CONTROLLER_LISTENER_NAMES=CONTROLLER
+      - KAFKA_CFG_LISTENERS=PLAINTEXT://:9092,CONTROLLER://:9093,EXTERNAL://:9094
+      - KAFKA_CFG_LISTENER_SECURITY_PROTOCOL_MAP=CONTROLLER:PLAINTEXT,PLAINTEXT:PLAINTEXT,EXTERNAL:PLAINTEXT
+      - KAFKA_CFG_ADVERTISED_LISTENERS=PLAINTEXT://kafka:9092,EXTERNAL://127.0.0.1:9094
+      - KAFKA_CFG_CONTROLLER_QUORUM_VOTERS=0@127.0.0.1:9093
+      - ALLOW_PLAINTEXT_LISTENER=yes
