--- src/main/kotlin/com/wypl/notification/emitter/NotificationServiceImpl.kt
@@ -1,77 +1,0 @@
-package com.wypl.notification.emitter
-
-import com.wypl.notification.global.message.NotificationSendable
-import com.wypl.notification.global.message.SubscribeMessage
-import com.wypl.notification.global.service.EmitterSubscribeService
-import io.github.oshai.kotlinlogging.KotlinLogging
-import org.springframework.stereotype.Service
-import org.springframework.web.servlet.mvc.method.annotation.SseEmitter
-import java.io.IOException
-
-private val logger = KotlinLogging.logger {}
-
-@Service
-class NotificationServiceImpl(
-    private val sseEmitterRepository: SseEmitterRepository
-) : EmitterSubscribeService {
-    companion object {
-        const val TIME_OUT_MILLISECONDS: Long = 60 * 1_000L
-    }
-
-    override fun connections(memberId: Int): SseEmitter {
-        logger.debug { "\n\n${memberId}번 사용자의 연결을 요청하였습니다.\n" }
-
-        val sseEmitter = SseEmitter(TIME_OUT_MILLISECONDS)
-        sseEmitterRepository.save(memberId, sseEmitter)
-
-        sseEmitterCallback(sseEmitter, memberId)
-
-        return sendMessage(
-            message = SubscribeMessage(
-                memberId = memberId,
-                timeout = TIME_OUT_MILLISECONDS
-            ),
-            memberId,
-            sseEmitter
-        )
-    }
-
-    private fun sseEmitterCallback(
-        sseEmitter: SseEmitter,
-        memberId: Int
-    ) {
-        sseEmitter.onCompletion {
-            logger.debug { "\n\n${memberId}번 사용자의 연결이 만료되어 삭제합니다.\n" }
-
-            sseEmitterRepository.deleteById(memberId)
-        }
-        sseEmitter.onTimeout {
-            logger.debug { "\n\n${memberId}번 사용자의 연결 요청이 Timeout으로 인해 종료합니다.\n" }
-
-            sseEmitter.complete()
-        }
-        sseEmitter.onError {
-            logger.debug { "\n\n${memberId}번 사용자의 연결 요청이 실패하여 종료합니다.\n" }
-
-            sseEmitter.complete()
-        }
-    }
-
-    fun sendMessage(
-        message: NotificationSendable,
-        memberId: Int,
-        sseEmitter: SseEmitter
-    ): SseEmitter {
-        try {
-            sseEmitter.send(
-                SseEmitter.event()
-                    .id(memberId.toString())
-                    .name(message.getMessageType().name)
-                    .data(message.getData())
-            )
-        } catch (e: IOException) {
-            sseEmitterRepository.deleteById(memberId)
-        }
-        return sseEmitter
-    }
-}
