+++ src/main/kotlin/com/wypl/notification/emitter/kafka/NotificationKafkaProducerMessageType.kt
@@ -1,0 +1,5 @@
+package com.wypl.notification.emitter.kafka
+
+enum class NotificationKafkaProducerMessageType {
+    SUCCESS_OF_FAILURE
+}
