+++ src/main/kotlin/com/wypl/notification/emitter/message/MessageType.kt
@@ -1,0 +1,5 @@
+package com.wypl.notification.emitter.message
+
+enum class MessageType {
+    SUBSCRIBE
+}
