+++ src/main/kotlin/com/wypl/notification/emitter/message/NotificationSendable.kt
@@ -1,0 +1,7 @@
+package com.wypl.notification.emitter.message
+
+interface NotificationSendable {
+    fun getMessageType(): MessageType
+
+    fun getData(): Map<String, Any>
+}
