+++ src/main/kotlin/com/wypl/notification/emitter/service/EmitterSubscribeService.kt
@@ -1,0 +1,3 @@
+package com.wypl.notification.emitter.service
+
+interface EmitterSubscribeService : SubscribeService
