--- src/main/kotlin/com/wypl/notification/global/config/embedded/EmbeddedMongoConfig.kt
@@ -1,12 +1,0 @@
-package com.wypl.notification.global.config.embedded
-
-import de.flapdoodle.embed.mongo.spring.autoconfigure.EmbeddedMongoAutoConfiguration
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration
-import org.springframework.context.annotation.Configuration
-import org.springframework.context.annotation.Profile
-
-@Profile("dev", "prod")
-@EnableAutoConfiguration(exclude = [EmbeddedMongoAutoConfiguration::class])
-@Configuration
-class EmbeddedMongoConfig {
-}
