--- src/main/kotlin/com/wypl/notification/global/config/embedded/redis/RedisAvailablePortFindable.kt
@@ -1,6 +1,0 @@
-package com.wypl.notification.global.config.embedded.redis
-
-abstract class RedisAvailablePortFindable {
-    abstract fun findAvailablePort(redisPort: Int): Int
-    abstract fun isRedisRunning(redisPort: Int): Boolean
-}
