--- src/main/kotlin/com/wypl/notification/global/controller/Subscribable.kt
@@ -1,8 +1,0 @@
-package com.wypl.notification.global.controller
-
-import org.springframework.http.ResponseEntity
-import org.springframework.web.servlet.mvc.method.annotation.SseEmitter
-
-interface Subscribable {
-    fun subscribe(memberId: Int): ResponseEntity<SseEmitter>
-}
