+++ src/main/kotlin/com/wypl/notification/global/kafka/KafkaConsumerConfig.kt
@@ -1,0 +1,31 @@
+package com.wypl.notification.global.kafka
+
+import org.apache.kafka.clients.consumer.ConsumerConfig
+import org.apache.kafka.common.serialization.StringDeserializer
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.kafka.annotation.EnableKafka
+import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory
+import org.springframework.kafka.core.ConsumerFactory
+import org.springframework.kafka.core.DefaultKafkaConsumerFactory
+
+@EnableKafka
+@Configuration
+class KafkaConsumerConfig {
+    @Bean
+    fun consumerFactory(): ConsumerFactory<String, Any> {
+        val properties: MutableMap<String, Any> = HashMap()
+        properties[ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG] = "127.0.0.1:9092"  // TODO: 수정
+        properties[ConsumerConfig.GROUP_ID_CONFIG] = "wypl-notification-group"  // TODO: 수정
+        properties[ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG] = StringDeserializer::class.java
+        properties[ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG] = StringDeserializer::class.java
+        return DefaultKafkaConsumerFactory(properties)
+    }
+
+    @Bean
+    fun kafkaListenerContainerFactory(): ConcurrentKafkaListenerContainerFactory<String, Any> {
+        val factory = ConcurrentKafkaListenerContainerFactory<String, Any>()
+        factory.consumerFactory = consumerFactory()
+        return factory
+    }
+}
