+++ src/main/kotlin/com/wypl/notification/global/kafka/KafkaMessage.kt
@@ -1,0 +1,3 @@
+package com.wypl.notification.global.kafka
+
+interface KafkaMessage
