--- src/main/kotlin/com/wypl/notification/global/message/MessageType.kt
@@ -1,5 +1,0 @@
-package com.wypl.notification.global.message
-
-enum class MessageType {
-    SUBSCRIBE
-}
