--- src/main/kotlin/com/wypl/notification/global/message/NotificationSendable.kt
@@ -1,7 +1,0 @@
-package com.wypl.notification.global.message
-
-interface NotificationSendable {
-    fun getMessageType(): MessageType
-
-    fun getData(): Map<String, Any>
-}
