+++ src/main/kotlin/com/wypl/notification/global/redis/embedded/OS.kt
@@ -1,0 +1,13 @@
+package com.wypl.notification.global.redis.embedded
+
+enum class OS(private val value: String) {
+    MAC("Mac"),
+    WINDOWS("Windows"),
+    LINUX("Linux"),
+    UBUNTU("Ubuntu"),
+    DEBIAN("Debian");
+
+    fun contains(osName: String): Boolean {
+        return osName.contains(value)
+    }
+}
