+++ src/main/kotlin/com/wypl/notification/global/redis/embedded/RedisAvailablePortFindable.kt
@@ -1,0 +1,6 @@
+package com.wypl.notification.global.redis.embedded
+
+abstract class RedisAvailablePortFindable {
+    abstract fun findAvailablePort(redisPort: Int): Int
+    abstract fun isRedisRunning(redisPort: Int): Boolean
+}
