--- src/main/kotlin/com/wypl/notification/global/service/EmitterSubscribeService.kt
@@ -1,3 +1,0 @@
-package com.wypl.notification.global.service
-
-interface EmitterSubscribeService : SubscribeService
