+++ src/main/kotlin/com/wypl/notification/schedule/message/ScheduleKafkaConsumerMessageType.kt
@@ -1,0 +1,5 @@
+package com.wypl.notification.schedule.message
+
+enum class ScheduleKafkaConsumerMessageType {
+    WRITE_REVIEW,
+}
