+++ src/main/resources/application.yml
@@ -14,13 +14,8 @@ spring:
     redis:
       port: 6379
       host: 127.0.0.1
-
 de:
   flapdoodle:
     mongodb:
       embedded:
         version: 4.0.2
-
-logging:
-  level:
-    root: debug
